Create webScraper.py
This was my first attempt at writing a webscraper in python. I learned a lot about the main webscraper libraries, Selenium and BeautifulSoup(not used) and while at times it was frustrating to get it working as intended due to the data changing across the pages, I thoroughly enjoyed the challenge and came away from it as a better programmer. The code will break on the last page so it isn't perfect, but it did deliver the results that I was after.