Add files via upload
This was a screening question asked of me in an interview for an open role. I am not very well experienced in front-end technologies but as always, I enjoy a good challenge. I learned a lot about javascript during this example and it was a good refresher into HTML elements that I haven't really dealt with since my college days. Overall, it was a moderately challenging question that I was happy to get working at the end.